+++ utility/src/oskar_Mem.cpp
@@ -37,7 +37,7 @@ oskar_Mem::oskar_Mem(int owner_)
     oskar_mem_init(this, 0, 0, 0, owner_, &err);
 }
 
-oskar_Mem::oskar_Mem(int mem_type, int mem_location, int size, int owner_)
+oskar_Mem::oskar_Mem(int mem_type, int mem_location, size_t size, int owner_)
 {
     int err = 0;
     oskar_mem_init(this, mem_type, mem_location, size, owner_, &err);
